requirements.txt
arrow==1.2.3
asgiref==3.7.2
attrs==23.1.0
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cookiecutter==2.3.0
Django==4.2.4
django-browser-reload==1.11.0
django-filter==23.3
django-tailwind==3.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.9.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.6.0
